refactor: :recycle: extracting the hook to a seperate file
