+++ src/pages/help-centre/index.tsx
@@ -12,7 +12,8 @@ import { SEO, Desktop, Container } from 'components/containers'
 import { Header } from 'components/elements'
 import Layout from 'components/layout/layout'
 import { localize, WithIntl } from 'components/localization'
-import { getLocationHash, sanitize, usePlatformQueryParam } from 'common/utility'
+import { getLocationHash, sanitize } from 'common/utility'
+import { usePlatformQueryParam } from 'components/hooks/use-platform-query-param'
 import device from 'themes/device'
 // Icons
 import SearchIcon from 'images/svg/help/search.svg'
